ALBS-952: Extract packages release with hawkey
Also, moved get_modules_yaml_from_repo into alws.utils.modularity
